requirements.txt

streamlit==1.8.0
tensorflow-cpu==2.8.0
click<=8.0.4
